refactor: hoist abort-aware sleep helper to apps/api/src/shared
The image-generation shim's rate-limit retry loop needed the same
setTimeout+AbortSignal wrapper that microsoft-grounding.ts already
carried verbatim. Extract a single shared `sleep(delayMs, signal)` and
import it from both call sites; delete the two local copies.
